ongoing error

+++ sql_scripts/func_procs_queries/updates.sql
@@ -205,8 +205,8 @@ DELIMITER &&
 CREATE PROCEDURE cancel_ongoing_sale (IN s_id INTEGER)
 BEGIN
 	DECLARE no_pds INT;
-    DECLARE pd_id INT;
-    DECLARE pd_quant_sold INT;
+	DECLARE pd_id INT;
+	DECLARE pd_quant_sold INT;
 	DECLARE check_error BOOLEAN DEFAULT FALSE;
     
 	DECLARE CONTINUE HANDLER FOR SQLEXCEPTION SET @check_error = TRUE;
@@ -221,22 +221,25 @@ BEGIN
 		REPEAT -- Selecionar produtos para adicionar no stock 
 			SELECT SP.ProductID_sp, SP.Quantity INTO pd_id, pd_quant_sold
 				FROM saleproduct AS SP 
-                INNER JOIN Product AS P 
-					ON SP.ProductID_sp = P.ProductID
+				INNER JOIN Product AS P 
+				ON SP.ProductID_sp = P.ProductID
 				WHERE SP.ReceiptNO_sp = s_id
-                ORDER BY SP.ProductID_sp DESC
-                LIMIT 1;
+				ORDER BY SP.ProductID_sp DESC
+				LIMIT 1;
                 
             -- atualizar stock do produto selecionado    
 			UPDATE Product
 			SET QuantityInStock = QuantityInStock + pd_quant_sold
-            WHERE ProductID = pd_id;
-            
-            -- retirar a venda do produto selecionado 
-            DELETE FROM SaleProduct
+			WHERE ProductID = pd_id;
+
+			IF check_error = FALSE THEN
+				-- retirar a venda do produto selecionado 
+				DELETE FROM SaleProduct
 				WHERE ReceiptNO_sp = s_id AND ProductID_sp = pd_id;
                 
-			SET no_pds = no_pds - 1;
+				SET no_pds = no_pds - 1;
+			END IF;
+
 		UNTIL no_pds <= 0 OR check_error != FALSE
 		END REPEAT;
